+++ changelog.md
@@ -1,3 +1,5 @@
+- <a href="https://github.com/groupby/issues/issues/920">iss14</a> Category expansion breaks Beautifier
+
 - <a href="https://github.com/groupby/issues/issues/739">iss13</a> URL Beautifier toUrlAdd function concatenating previous refinements
 
 - <a href="https://github.com/groupby/issues/issues/720">iss11</a> Quickstart to support more refinements